requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-environ==0.12.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
